training duration

+++ cpp/core/common.cpp
@@ -6,7 +6,7 @@ using namespace std;
 MPI_Datatype distblas::core::SPTUPLE;
 MPI_Datatype distblas::core::DENSETUPLE;
 
-uint64_t distblas::core::divide_and_round_up(uint64_t num, int denom){
+int distblas::core::divide_and_round_up(uint64_t num, int denom){
   if (num % denom > 0) {
     return num / denom + 1;
   }
